Djinn.GHC

Documentation

type Environment = [(Name, Type)]

Bindings which are in scope at a specific point.

newtype MaxSolutions

Obtain a maximum number of solutions.

Constructors

Max Int

djinn :: GhcMonad m => Bool -> Maybe ModuleInfo -> Environment -> Type -> MaxSolutions -> Int -> m [String]

Obtain the list of expressions which could fill something with the given type. The first flag specifies whether to return one or more solutions to the problem.